pyproject.toml
[tool.poetry]
name = "pylsr"
version = "2020"
description = "Use this module to read, and write to Apple's Layer Source Representation (LSR) Image format."
authors = ["FredHappyface"]
classifiers = [
	"Programming Language :: Python :: 3.6",
	"Programming Language :: Python :: 3.7",
	"Programming Language :: Python :: 3.8",
	"Operating System :: OS Independent"
]
homepage = "https://github.com/FHPythonUtils/PyLSR"
repository = "https://github.com/FHPythonUtils/PyLSR"
documentation = "https://github.com/FHPythonUtils/PyLSR/blob/master/README.md"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.5"
pillow = "^7.1.2"


[tool.dephell.main]
from = "pyproject.toml"
to = "requirements.txt"
envs = ["main"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
